Remove scaffolding
